Use Preprocessor from interface module
