+++ amm/interp/src/main/scala/ammonite/interp/Interpreter.scala
@@ -4,7 +4,7 @@ import java.io.{File, OutputStream, PrintStream}
 import java.util.regex.Pattern
 
 import ammonite.compiler._
-import ammonite.compiler.iface.CodeWrapper
+import ammonite.compiler.iface.{CodeWrapper, Preprocessor}
 import ammonite.interp.api.{InterpAPI, InterpLoad, LoadJar}
 
 import scala.collection.mutable
@@ -377,7 +377,7 @@ class Interpreter(val printer: Printer,
         // and none of it's blocks end up needing to be re-compiled. We don't know up
         // front if any blocks will need re-compilation, because it may import $file
         // another script which gets changed, and we'd only know when we reach that block
-        lazy val splittedScript = Preprocessor.splitScript(
+        lazy val splittedScript = Parsers.splitScript(
           Interpreter.skipSheBangLine(code),
           codeSource.fileName
         )
@@ -422,7 +422,7 @@ class Interpreter(val printer: Printer,
     val wrapperName = Name("cmd" + currentLine)
     val fileName = wrapperName.encoded + ".sc"
     for {
-      blocks <- Res(Preprocessor.splitScript(Interpreter.skipSheBangLine(code), fileName))
+      blocks <- Res(Parsers.splitScript(Interpreter.skipSheBangLine(code), fileName))
 
       metadata <- processAllScriptBlocks(
         blocks.map(_ => None),
